+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Bump Mobius to v1.5.6
 - Enable overdue list download and share feature in India only
 - Set user ID as user property in `MixpanelAnalyticsReporter`
+- Extend `RxWorker` in `OverdueDownloadWorker`
 
 ### Changes